fixed duplicate assignemnt submissions
